Implement check for Permission to Comment in ReportbookService  (#148)
* Add check for permission to editComment

* Add check for permission to deleteComment

+++ src/Reportbook/ReportbookService.php
@@ -162,17 +162,23 @@ class ReportbookService
      * @param string $id
      * @return Comment
      */
-    public function editComment(string $id, string $newContent): Comment
+    public function editComment(string $id, string $newContent, string $userId): Comment
     {
-        return $this->commentService->editComment($id, $newContent);
+        $comment = $this->findCommentById($id);
+        if ($userId === $comment->userId()) {
+            return $this->commentService->editComment($id, $newContent);
+        }
     }
 
     /**
      * @param string $commentId
      */
-    public function deleteComment(string $commentId)
+    public function deleteComment(string $commentId, string $userId)
     {
-        $this->commentService->deleteComment($commentId);
+        $comment = $this->findCommentById($commentId);
+        if ($userId === $comment->userId()) {
+            $this->commentService->deleteComment($commentId);
+        }
     }
 
     /**
